% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classify.dd.R
\name{feDP}
\alias{feDP}
\title{feDP}
\usage{
feDP(pe_mat, condition, sig_genes, oa, c1, c2, log.nonzero = TRUE)
}
\arguments{
\item{pe_mat}{Matrix with genes in rows and samples in columns.  Column names
 indicate condition.}

\item{condition}{Vector of condition indicators, where each item is equal to 1 or 2.}

\item{sig_genes}{Vector of the indices of significantly DD genes (indicating the row
 number of \code{pe_mat})}

\item{oa}{List item with one item for each gene where the first element contains the cluster membership for
 each nonzero sample in the overall (pooled) fit.}

\item{c1}{List item with one item for each gene where the first element contains the cluster membership for
 each nonzero sample in condition 1 only fit}

\item{c2}{List item with one item for each gene where the first element contains the cluster membership for
 each nonzero sample in condition 2 only fit}

\item{log.nonzero}{Logical indicating whether to perform log transformation of nonzero values.}
}
\value{
cat Character vector of the same length as \code{sig_genes} that indicates which nonsignificant genes by
 the permutation test belong to the DP category
}
\description{
Function to identify additional DP genes, since clustering process can be consistent within each condition
  and still have differential proportion within each mode.
}
\details{
The Fisher's Exact test is used to test for independence of condition membership and clustering when
  the clustering is the same within condition as it is overall (and is multimodal)
}

