% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster_actions.R
\name{lu}
\alias{lu}
\title{lu}
\usage{
lu(x)
}
\arguments{
\item{x}{Numeric vector of cluster membership (1st item (named \code{class}) in list returned by \code{\link{mclustRestricted}})}
}
\value{
The count of the number of unique clusters.
}
\description{
Shortcut for \code{length(unique())}
}
\details{
Function to obtain a count of the number of clusters
}

