% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust.restricted.R
\name{mclustRestricted}
\alias{mclustRestricted}
\title{mclustRestricted}
\usage{
mclustRestricted(y, restrict = TRUE)
}
\arguments{
\item{y}{Numeric vector of values to fit to a normal mixture model with Mclust.}

\item{restrict}{Logical indicating whether or not to enforce the restriction on cluster separation by at least 2 standard deviations
and ratio of largest to smallest variance less than 3 when small clusters are present (containing 10% or less of the sample).
If False, then Mclust results as is are returned.}
}
\value{
List object with (1) vector of cluster membership, (2) cluster means, (3) cluster variances, (4) number of model parameters,
 (5) sample size, (6) BIC of selected model, and (6) loglikelihood of selected model. 
 @export
}
\description{
Function to determine how many normal mixture components are present.
}
\details{
Robust to detecting multiple components that are close together by enforcing that the distance between two clusters 
 of appreciable size (at least 4 samples) be at least 2 standard deviations and not have variances that differ by more than 
 a ratio of 3 for clusters of small size (10% or less of the sample).
}

