% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{normExprs}
\alias{normExprs}
\title{normExprs}
\usage{
normExprs(SCdat)
}
\arguments{
\item{SCdat}{An object of class \code{SummarizedExperiment} that contains 
normalized single-cell expression and metadata}
}
\value{
A \code{matrix} which contains the normalized count data
 where genes are in rows and cells are in columns
}
\description{
extract Normalized expression matrix from SummarizedExperiment object
}
\details{
Convenient helper function to extract the normalized expression
matrix (in the \code{"NormedCounts"} slot of \code{assays}) from the
SummarizedExperiment
}
\examples{
 
# load toy simulated example SummarizedExperiment to find DD genes  
data(scDatExSim)

# extract normalized expression matrix from the SummarizedExperiment object 

normCounts <- normExprs(scDatExSim)
}
