% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permMclustCov}
\alias{permMclustCov}
\title{permMclustCov}
\usage{
permMclustCov(y, nperms, C, condition, remove.zeroes = TRUE,
  log.transf = TRUE, restrict = FALSE, alpha, m0, s0, a0, b0)
}
\arguments{
\item{y}{Numeric data vector for one gene}

\item{nperms}{Number of permutations of residuals to evaulate}

\item{C}{Matrix of confounder variables, where there is one row for each sample and one column for each covariate.}

\item{condition}{Vector of condition indicators for each sample}

\item{remove.zeroes}{Logical indicating whether zeroes need to be removed from \code{y}}

\item{log.transf}{Logical indicating whether the data is in the raw scale (if so, will be log-transformed)}

\item{restrict}{Logical indicating whether to perform restricted Mclust clustering where close-together clusters are joined.}

\item{alpha}{Value for the Dirichlet concentration parameter}

\item{m0}{Prior mean value for generating distribution of cluster means}

\item{s0}{Prior precision value for generating distribution of cluster means}

\item{a0}{Prior shape parameter value for the generating distribution of cluster precision}

\item{b0}{Prior scale parameter value for the generating distribution of cluster precision}
}
\value{
Bayes factor numerator for the current permutation
}
\description{
Function to  obtain bayes factor for permutations of one gene's residuals
}
\details{
Obtains bayes factor numerator for data vector \code{y} representing one gene where \code{p1} denotes which
}

