% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ex-data.R
\docType{data}
\name{scDatExList}
\alias{scDatExList}
\title{Toy example data list}
\format{A list of two matrices (one for each of two conditions) labeled "C1" and "C2".  Each matrix contains data for 100 genes
  and a variable number of samples (78 in C1 and 64 in C2).}
\usage{
data(scDatExList)
}
\description{
Toy example data list (one item for each of two conditions) for 100 genes to illustrate how to use the function
 \code{\link{preprocess}}.
}
\references{
Subset of data from the H1 and DEC comparison in the Korthauer et al. (2015) paper (In preparation).
}
\keyword{datasets}

