% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDD.R
\name{scDD}
\alias{scDD}
\title{scDD}
\usage{
scDD(SCdat, prior_param = list(alpha = 0.1, mu0 = 0, s0 = 0.01, a0 = 0.01, b0
  = 0.01), permutations = 100, testZeroes = TRUE, adjust.perms = FALSE)
}
\arguments{
\item{SCdat}{An object of class \code{ExpressionSet} that contains normalized single-cell expression and metadata, where the \code{assayData} 
slot contains one row for each gene and one sample for each column.  The \code{PhenoData} slot should contain a vector of numeric values
(either 1 or 2) that indicates which 
condition each sample belongs to (in the same order as the columns of \code{assayData}).  Optional additional metadata about the 
experiment can be contained in the \code{experimentData} slot.}

\item{prior_param}{A list of prior parameter values to be used when modeling each gene as a mixture of DP normals.  Default 
values are given that specify a vague prior distribution on the cluster-specific means and variances.}

\item{permutations}{The number of permutations to be used in calculating empirical p-values.}

\item{testZeroes}{Logical indicating whether or not to test for a difference in the proportion of zeroes}

\item{adjust.perms}{Logical indicating whether or not to adjust the permutation tests for the sample
detection rate (proportion of nonzero values).  If true, the residuals of a linear model adjusted for 
detection rate are permuted, and new fitted values are obtained using these residuals.}
}
\value{
List with four items: the first is a data frame with seven columns: gene name (matches rownames of SCdat), permutation p-value for testing of independence of 
 condition membership with clustering, p-value for test of difference in dropout rate (only for non-DD genes), and name of the 
 DD (DE, DP, DM, DB) pattern or DZ (otherwise NS = not significant), the number of clusters identified overall, the number of clusters identified in 
 condition 1 alone, and the number of clusters identified in condition 2 alone.    The remaining three elements are data frames (first for condition 1 and 2 combined, then condition 1 alone, then condition 2 alone) that contains the cluster memberships for each sample (cluster 1,2,3,...) in columns and genes in rows.  Zeroes, which are not involved in the clustering, are labeled as zero.
}
\description{
Find genes with differential distributions (DD) across two conditions
}
\details{
Find genes with differential distributions (DD) across two conditions.  Models each log-transformed gene as a Dirichlet 
  Process Mixture of normals and uses a permutation test to determine whether condition membership is independent of sample clustering.
  The FDR adjusted (Benjamini-Hochberg) permutation p-value is returned along with the classification of each significant gene 
  (with p-value less than 0.05 (or 0.025 if also testing for a difference in the proportion of zeroes)) into one of four categories 
  (DE, DP, DM, DB).  For genes that do not show significant influence, of condition on clustering, an optional test of whether the 
  proportion of zeroes (dropout rate) is different across conditions is performed (DZ).
}
\examples{
 
 # load toy simulated example ExpressionSet to find DD genes

data(scDatExSim)


# check that this object is a member of the ExpressionSet class
# and that it contains 200 samples and 30 genes

class(scDatExSim)
show(scDatExSim)


# set arguments to pass to scDD function
# we will perform 100 permutations on each of the 30 genes

prior_param=list(alpha=0.01, mu0=0, s0=0.01, a0=0.01, b0=0.01)
nperms <- 100


# call the scDD function to perform permutations, classify DD genes, and return results
# we won't perform the test for a difference in the proportion of zeroes since none 
# exists in this simulated toy example data
# this step will take significantly longer with more genes and/or more permutations

RES <- scDD(scDatExSim, prior_param=prior_param, permutations=nperms, testZeroes=FALSE)
}

