% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Cell_Simu.R
\name{calcRP}
\alias{calcRP}
\title{calcRP}
\usage{
calcRP(Emean, Evar)
}
\arguments{
\item{Emean}{Empirical mean}

\item{Evar}{Empirical variance}
}
\value{
RP Vector of two elements, first contains method of moments estimator for r and second contains 
  method of moments estimator for p (parameters of NB distribution)
}
\description{
Calculate parameter R and P in NB distribution
}
\references{
Korthauer KD, Chu LF, Newton MA, Li Y, Thomson J, Stewart R, Kendziorski C. A statistical approach for identifying differential distributions
in single-cell RNA-seq experiments. Genome Biology. 2016 Oct 25;17(1):222. \url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-016-1077-y}
}

