% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIndex.R
\name{findIndex}
\alias{findIndex}
\title{findIndex}
\usage{
findIndex(SCdat)
}
\arguments{
\item{SCdat}{An object of class \code{ExpressionSet} that contains normalized single-cell expression and metadata, where the \code{assayData} 
slot contains one row for each gene and one sample for each column.  The \code{PhenoData} slot should contain a vector of numeric values
(either 1 or 2) that indicates which 
condition each sample belongs to (in the same order as the columns of \code{assayData}).  Optional additional metadata about the 
experiment can be contained in the \code{experimentData} slot.}
}
\value{
Vector of indices for a reasonable set of genes that can be used for simulation.
}
\description{
Find a reasonable set of genes (one mode and at least 25% nonzero values) to use for simulation.
}

