% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/joint.posterior.R
\name{jointPosterior}
\alias{jointPosterior}
\title{jointPosterior}
\usage{
jointPosterior(y, mcobj, alpha, m0, s0, a0, b0)
}
\arguments{
\item{y}{Numeric data vector for one gene (log-transformed non-zeroes)}

\item{mcobj}{Object returned by \code{\link{mclustRestricted}}}

\item{alpha}{Value for the Dirichlet concentration parameter}

\item{m0}{Prior mean value for generating distribution of cluster means}

\item{s0}{Prior precision value for generating distribution of cluster means}

\item{a0}{Prior shape parameter value for the generating distribution of cluster precision}

\item{b0}{Prior scale parameter value for the generating distribution of cluster precision}
}
\value{
log joint posterior value
}
\description{
Function to obtain the normalized joint posterior of the data and partition.
}
\details{
Calculates the normalized joint posterior of the data and partition under the Product Partition Model formulation
  of the Dirichlet Process Mixture model.
}

