% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_actions.R
\name{lu}
\alias{lu}
\title{lu}
\usage{
lu(x)
}
\arguments{
\item{x}{Numeric vector of cluster membership (1st item (named \code{class})
in list returned by \code{\link{mclustRestricted}})}
}
\value{
The count of the number of unique clusters.
}
\description{
Shortcut for \code{length(unique())}
}
\details{
Function to obtain a count of the number of clusters
}
\references{
Korthauer KD, Chu LF, Newton MA, Li Y, Thomson J, Stewart R, 
Kendziorski C. A statistical approach for identifying differential 
distributions
in single-cell RNA-seq experiments. Genome Biology. 2016 Oct 25;17(1):222. 
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-016-
1077-y}
}

