% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(DataList, ConditionNames, zero.thresh = 0.9, median_norm = FALSE)
}
\arguments{
\item{DataList}{A list object where each item contains a matrix of data with rows 
designating genes and samples designating columns.  The name of the list objects represents
their condition.}

\item{ConditionNames}{Character vector of length 1 or 2 which contains the name(s) of the
conditions (the names of the items in \code{DataList}) to be processed.}

\item{zero.thresh}{A numeric value between 0 and 1 that represents the maximum proportion
of zeroes per gene allowable in the processed dataset}

\item{median_norm}{Logical indicating whether or not to normalize the data using Median Normalization from \code{EBSeq}}
}
\value{
pe_mat Processed matrix with genes in rows and samples in columns.  Column names 
 indicate condition.
 
 @examples 
 
 # load toy example data list
 
 data(scDatExList)
 
 
 # check that the data is formated as a list of 2 matrices (one for each of 2 conditions), 
 # that each matrix has 100 rows (one for each gene), and that the number of columns in 
 # each matrix corresponds to the number of samples in each condition (78 and 64, respectively)
 
 str(scDatExList)
 
 
 # get the names of the conditions to pass to the preprocess function
 
 condition.names <- names(scDatExList)
 
 
 # apply the preprocess function to reformat the data into one data matrix with 100 rows and
 # 78+64=142 columns
 # set the zero.thresh argument to 1 so that genes are filtered out if they are all zero
 # set the median_norm argument to FALSE to return raw data
 
 scDatExMat <- preprocess(scDatExList, ConditionNames=condition.names, 
                          zero.thresh=1, median_norm=FALSE)
 
 
 # apply the preprocess function again, but this time threshold on the proportion of zeroes 
 # and apply median normalization
 # set the zero.thresh argument to 0.75 so that genes with more than 75% zeroes are filtered out
 # set the median_norm argument to TRUE to return median normalized counts
 
 scDatExMatNormThresh <- preprocess(scDatExList, ConditionNames=condition.names, 
                                    zero.thresh=0.75, median_norm=TRUE)
}
\description{
Function to preprocess a list of data matrices to (1) combine into one matrix and (2)
 only keep genes with a certain number of nonzero entries.
}
\references{
Korthauer KD, Chu LF, Newton MA, Li Y, Thomson J, Stewart R, Kendziorski C. A statistical approach for identifying differential distributions
in single-cell RNA-seq experiments. Genome Biology. 2016 Oct 25;17(1):222. \url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-016-1077-y}
}

