% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ex-data.R
\docType{data}
\name{scDatEx}
\alias{scDatEx}
\title{Toy example data}
\format{An object of class \code{\link[Biobase]{ExpressionSet}} containing data for 500 genes for 142 samples
 (78 from condition 1 and 64 from condition 2).  Condition labels (1 or 2) are stored in the phenoData slot.}
\usage{
data(scDatEx)
}
\description{
Toy example data in \code{\link[Biobase]{ExpressionSet}} format for 500 genes to illustrate how to generate simulated data from example data
 using \code{\link{simulateSet}}.
}
\references{
Subset of data from the H1 and DEC comparison in the Korthauer et al. (2015) paper (In preparation).
}
\keyword{datasets}

