% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SidebySideHists.R
\name{sideHist}
\alias{sideHist}
\title{sideHist}
\usage{
sideHist(x, y, logT = TRUE, title.gene = "")
}
\arguments{
\item{x}{First numeric vector of data to plot.}

\item{y}{Second numeric vector of data to plot.}

\item{logT}{Logical that indicates whether to take the log(x+1) transformation.}

\item{title.gene}{Character vector that contains the gene name that you are plotting}
}
\description{
Plots two histograms side by side with smoothed density overlay
}

