% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{validation}
\alias{validation}
\title{validation}
\usage{
validation(MV, DEIndex, Zeropercent_Base, Simulated_Data, numGenes)
}
\arguments{
\item{MV}{Mean and Variance matrix for observed data}

\item{DEIndex}{Index for genes chosen to be DE (can be NULL)}

\item{Zeropercent_Base}{Zero percentage for corresponding gene expression 
values}

\item{Simulated_Data}{Simulated dataset}

\item{numGenes}{numeric value for the number of genes to simulate}
}
\value{
Validation plots
}
\description{
Draw validation plots to show that the simulated dataset emulates 
characteristics of observed dataset.
}
\references{
Korthauer KD, Chu LF, Newton MA, Li Y, Thomson J, Stewart R, 
Kendziorski C. A statistical approach for identifying differential 
distributions
in single-cell RNA-seq experiments. Genome Biology. 2016 Oct 25;17(1):222. 
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-016-
1077-y}
}

